+++ source/Private/Get-BundledToolInfo.ps1
@@ -10,6 +10,7 @@ function Get-BundledToolInfo {
         [PSCustomObject]@{ FileName = 'delphi-incver.ps1';  SupportsVersionApi = $false }
         [PSCustomObject]@{ FileName = 'delphi-coverage.ps1'; SupportsVersionApi = $true }
         [PSCustomObject]@{ FileName = 'delphi-callgraph.ps1'; SupportsVersionApi = $true }
+        [PSCustomObject]@{ FileName = 'delphi-codesign-azure.ps1'; SupportsVersionApi = $true }
     )
 
     $results = [System.Collections.Generic.List[object]]::new()